Serve frontend app with data app
