+++ data/service/helpers/exceptions/__init__.py
@@ -12,3 +12,4 @@ from data.service.helpers.exceptions.strategy_required import StrategyRequired
 from data.service.helpers.exceptions.symbol_required import SymbolRequired
 from data.service.helpers.exceptions.leverage_invalid import LeverageInvalid
 from data.service.helpers.exceptions.pipeline_start_fail import PipelineStartFail
+from data.service.helpers.exceptions.data_pipeline_does_not_exist import DataPipelineDoesNotExist
